KData functions refactored into Mixin classes
